fix: rename redirect
